Fill in hint precisions in place_hints function
