+++ randovania/generator/filler/runner.py
@@ -55,59 +55,6 @@ def _create_weighted_list(rng: Random,
     return current
 
 
-def add_hints_precision(patches: GamePatches,
-                        rng: Random,
-                        ) -> GamePatches:
-    """
-    Adds precision to all hints that are missing one.
-    :param patches:
-    :param rng:
-    :return:
-    """
-
-    hints_to_replace = {
-        asset: hint
-        for asset, hint in patches.hints.items()
-        if hint.precision is None
-    }
-
-    asset_ids = list(hints_to_replace.keys())
-
-    # Add random precisions
-    rng.shuffle(asset_ids)
-    precisions = []
-    for asset_id in asset_ids:
-        precisions = _create_weighted_list(rng, precisions, PrecisionPair.weighted_list)
-        precision = precisions.pop()
-
-        hints_to_replace[asset_id] = dataclasses.replace(hints_to_replace[asset_id], precision=precision)
-
-    # Replace the hints the in the patches
-    return dataclasses.replace(patches, hints={
-        asset: hints_to_replace.get(asset, hint)
-        for asset, hint in patches.hints.items()
-    })
-
-
-def replace_hints_without_precision_with_jokes(patches: GamePatches) -> GamePatches:
-    """
-    Adds WRONG_GAME precision to all hints that are missing one precision.
-    :param patches:
-    :return:
-    """
-
-    hints_to_replace = {
-        asset: dataclasses.replace(hint, precision=PrecisionPair.joke())
-        for asset, hint in patches.hints.items()
-        if hint.precision
-    }
-
-    return dataclasses.replace(patches, hints={
-        asset: hints_to_replace.get(asset, hint)
-        for asset, hint in patches.hints.items()
-    })
-
-
 def run_filler(configuration: LayoutConfiguration,
                game: GameDescription,
                item_pool: List[PickupEntry],
@@ -145,11 +92,4 @@ def run_filler(configuration: LayoutConfiguration,
         status_update=status_update)
 
     # Since we haven't added expansions yet, these hints will always be for items added by the filler.
-    full_hints_patches = place_hints(final_state, rng, game.world_list, status_update)
-
-    if configuration.hints.item_hints:
-        result = add_hints_precision(full_hints_patches, rng)
-    else:
-        result = replace_hints_without_precision_with_jokes(full_hints_patches)
-
-    return result, major_items + expansions
+    return place_hints(configuration, final_state, rng, game.world_list, status_update), major_items + expansions
